@@ -1,6 +1,5 @@
 import json
 from packaging.version import Version
-from typing import Dict, Optional, Tuple
 
 import numpy as np
 import pandas as pd
@@ -138,9 +137,6 @@ def geopandas_to_arrow(
     geometry_encoding_dict = {}
 
     if geometry_encoding.lower() == "geoarrow":
-        if Version(pa.__version__) < Version("10.0.0"):
-            raise ValueError("Converting to 'geoarrow' requires pyarrow >= 10.0.")
-
         # Encode all geometry columns to GeoArrow
         for i, col in zip(geometry_indices, geometry_columns):
             field, geom_arr = construct_geometry_array(
@@ -177,9 +173,8 @@ def construct_wkb_array(
     shapely_arr: NDArray[np.object_],
     *,
     field_name: str = "geometry",
-    crs: Optional[str] = None,
-) -> Tuple[pa.Field, pa.Array]:
-
+    crs: str | None = None,
+) -> tuple[pa.Field, pa.Array]:
     if shapely.geos_version > (3, 10, 0):
         kwargs = {"flavor": "iso"}
     else:
@@ -191,7 +186,7 @@ def construct_wkb_array(
     extension_metadata = {"ARROW:extension:name": "geoarrow.wkb"}
     if crs is not None:
         extension_metadata["ARROW:extension:metadata"] = json.dumps(
-            {"crs": crs.to_json()}
+            {"crs": crs.to_json_dict()}
         )
     else:
         # In theory this should not be needed, but otherwise pyarrow < 17
@@ -270,20 +265,27 @@ def _multipolygon_type(point_type):
 
 def construct_geometry_array(
     shapely_arr: NDArray[np.object_],
-    include_z: Optional[bool] = None,
+    include_z: bool | None = None,
     *,
     field_name: str = "geometry",
-    crs: Optional[str] = None,
+    crs: str | None = None,
     interleaved: bool = True,
-) -> Tuple[pa.Field, pa.Array]:
+) -> tuple[pa.Field, pa.Array]:
     # NOTE: this implementation returns a (field, array) pair so that it can set the
     # extension metadata on the field without instantiating extension types into the
     # global pyarrow registry
+
+    mask = shapely.is_missing(shapely_arr)
+    if len(shapely_arr) == 0 or mask.all():
+        raise NotImplementedError(
+            "Converting an empty or all-missing GeoDataFrame to the 'geoarrow' "
+            "encoding is not yet supported."
+        )
+
     geom_type, coords, offsets = shapely.to_ragged_array(
         shapely_arr, include_z=include_z
     )
 
-    mask = shapely.is_missing(shapely_arr)
     if mask.any():
         if (
             geom_type == GeometryType.POINT
@@ -314,10 +316,10 @@ def construct_geometry_array(
     else:
         raise ValueError(f"Unexpected coords dimensions: {coords.shape}")
 
-    extension_metadata: Dict[str, str] = {}
+    extension_metadata: dict[str, str] = {}
     if crs is not None:
         extension_metadata["ARROW:extension:metadata"] = json.dumps(
-            {"crs": crs.to_json()}
+            {"crs": crs.to_json_dict()}
         )
     else:
         # In theory this should not be needed, but otherwise pyarrow < 17
@@ -447,7 +449,7 @@ def _get_arrow_geometry_field(field):
     return None
 
 
-def arrow_to_geopandas(table, geometry=None):
+def arrow_to_geopandas(table, geometry=None, to_pandas_kwargs=None):
     """
     Convert Arrow table object to a GeoDataFrame based on GeoArrow extension types.
 
@@ -458,6 +460,12 @@ def arrow_to_geopandas(table, geometry=None):
     geometry : str, default None
         The name of the geometry column to set as the active geometry
         column. If None, the first geometry column found will be used.
+    to_pandas_kwargs : dict, optional
+        Arguments passed to the `pa.Table.to_pandas` method for non-geometry columns.
+        This can be used to control the behavior of the conversion of the non-geometry
+        columns to a pandas DataFrame. For example, you can use this to control the
+        dtype conversion of the columns. By default, the `to_pandas` method is called
+        with no additional arguments.
 
     Returns
     -------
@@ -478,7 +486,9 @@ def arrow_to_geopandas(table, geometry=None):
         raise ValueError("No geometry column found in the Arrow table.")
 
     table_attr = table.drop([f[1] for f in geom_fields])
-    df = table_attr.to_pandas()
+    if to_pandas_kwargs is None:
+        to_pandas_kwargs = {}
+    df = table_attr.to_pandas(**to_pandas_kwargs)
 
     for i, col, ext_name, ext_meta in geom_fields:
         crs = None
@@ -488,7 +498,6 @@ def arrow_to_geopandas(table, geometry=None):
         if ext_name == "geoarrow.wkb":
             geom_arr = from_wkb(np.array(table[col]), crs=crs)
         elif ext_name.split(".")[1] in GEOARROW_ENCODINGS:
-
             geom_arr = from_shapely(
                 construct_shapely_array(table[col].combine_chunks(), ext_name), crs=crs
             )
@@ -526,7 +535,6 @@ def arrow_to_geometry_array(arr):
     if ext_name == "geoarrow.wkb":
         geom_arr = from_wkb(np.array(pa_arr), crs=crs)
     elif ext_name.split(".")[1] in GEOARROW_ENCODINGS:
-
         geom_arr = from_shapely(construct_shapely_array(pa_arr, ext_name), crs=crs)
     else:
         raise ValueError(f"Unknown GeoArrow extension type: {ext_name}")
