@@ -1,6 +1,7 @@
 import json
 import warnings
 from packaging.version import Version
+from typing import Literal, get_args
 
 import numpy as np
 from pandas import DataFrame, Series
@@ -15,7 +16,8 @@ from geopandas.array import from_shapely, from_wkb
 from .file import _expand_user
 
 METADATA_VERSION = "1.0.0"
-SUPPORTED_VERSIONS = ["0.1.0", "0.4.0", "1.0.0-beta.1", "1.0.0", "1.1.0"]
+SUPPORTED_VERSIONS_LITERAL = Literal["0.1.0", "0.4.0", "1.0.0-beta.1", "1.0.0", "1.1.0"]
+SUPPORTED_VERSIONS = list(get_args(SUPPORTED_VERSIONS_LITERAL))
 GEOARROW_ENCODINGS = [
     "point",
     "linestring",
@@ -25,6 +27,7 @@ GEOARROW_ENCODINGS = [
     "multipolygon",
 ]
 SUPPORTED_ENCODINGS = ["WKB"] + GEOARROW_ENCODINGS
+PARQUET_GEOMETRY_ENCODINGS = Literal["WKB", "geoarrow"]
 
 # reference: https://github.com/opengeospatial/geoparquet
 
@@ -119,9 +122,12 @@ def _create_metadata(
     Parameters
     ----------
     df : GeoDataFrame
-    schema_version : {'0.1.0', '0.4.0', '1.0.0-beta.1', '1.0.0', None}
+    schema_version : {'0.1.0', '0.4.0', '1.0.0-beta.1', '1.0.0', '1.1.0', None}
         GeoParquet specification version; if not provided will default to
         latest supported version.
+    geometry_encoding : dict, default None
+        GeoParquet encoding per geometry column.
+        Defaults to "WKB" for columns that are not present in the dictionary.
     write_covering_bbox : bool, default False
         Writes the bounding box column for each row entry with column
         name 'bbox'. Writing a bbox column can be computationally
@@ -166,7 +172,11 @@ def _create_metadata(
                 _remove_id_from_member_of_ensembles(crs)
 
         column_metadata[col] = {
-            "encoding": geometry_encoding[col],
+            "encoding": (
+                geometry_encoding[col]
+                if geometry_encoding and col in geometry_encoding
+                else "WKB"
+            ),
             "crs": crs,
             geometry_types_name: geometry_types,
         }
@@ -271,8 +281,7 @@ def _validate_geo_metadata(metadata):
     version = metadata.get("version", metadata.get("schema_version"))
     if not version:
         raise ValueError(
-            "'geo' metadata in Parquet/Feather file is missing required key: "
-            "'version'"
+            "'geo' metadata in Parquet/Feather file is missing required key: 'version'"
         )
 
     required_keys = ("primary_column", "columns")
@@ -280,7 +289,7 @@ def _validate_geo_metadata(metadata):
         if metadata.get(key, None) is None:
             raise ValueError(
                 "'geo' metadata in Parquet/Feather file is missing required key: "
-                "'{key}'".format(key=key)
+                f"'{key}'"
             )
 
     if not isinstance(metadata["columns"], dict):
@@ -294,7 +303,7 @@ def _validate_geo_metadata(metadata):
             if key not in column_metadata:
                 raise ValueError(
                     "'geo' metadata in Parquet/Feather file is missing required key "
-                    "'{key}' for column '{col}'".format(key=key, col=col)
+                    f"'{key}' for column '{col}'"
                 )
 
         if column_metadata["encoding"] not in SUPPORTED_ENCODINGS:
@@ -414,7 +423,7 @@ def _to_parquet(
         The encoding to use for the geometry columns. Defaults to "WKB"
         for maximum interoperability. Specify "geoarrow" to use one of the
         native GeoArrow-based single-geometry type encodings.
-    schema_version : {'0.1.0', '0.4.0', '1.0.0', None}
+    schema_version : {'0.1.0', '0.4.0', '1.0.0', '1.1.0', None}
         GeoParquet specification version; if not provided will default to
         latest supported version.
     write_covering_bbox : bool, default False
@@ -465,7 +474,7 @@ def _to_feather(df, path, index=None, compression=None, schema_version=None, **k
     compression : {'zstd', 'lz4', 'uncompressed'}, optional
         Name of the compression to use. Use ``"uncompressed"`` for no
         compression. By default uses LZ4 if available, otherwise uncompressed.
-    schema_version : {'0.1.0', '0.4.0', '1.0.0', None}
+    schema_version : {'0.1.0', '0.4.0', '1.0.0', '1.1.0', None}
         GeoParquet specification version for the metadata; if not provided
         will default to latest supported version.
     kwargs
@@ -474,24 +483,19 @@ def _to_feather(df, path, index=None, compression=None, schema_version=None, **k
     feather = import_optional_dependency(
         "pyarrow.feather", extra="pyarrow is required for Feather support."
     )
-    # TODO move this into `import_optional_dependency`
-    import pyarrow
-
-    if Version(pyarrow.__version__) < Version("0.17.0"):
-        raise ImportError("pyarrow >= 0.17 required for Feather support")
 
     path = _expand_user(path)
     table = _geopandas_to_arrow(df, index=index, schema_version=schema_version)
     feather.write_feather(table, path, compression=compression, **kwargs)
 
 
-def _arrow_to_geopandas(table, geo_metadata=None):
+def _arrow_to_geopandas(table, geo_metadata=None, to_pandas_kwargs=None):
     """
     Helper function with main, shared logic for read_parquet/read_feather.
     """
-    geo_metadata = geo_metadata or _decode_metadata(
-        table.schema.metadata.get(b"geo", b"")
-    )
+    if geo_metadata is None:
+        # Note: this path of not passing metadata is also used by dask-geopandas
+        geo_metadata = _validate_and_decode_metadata(table.schema.metadata)
 
     # Find all geometry columns that were read from the file.  May
     # be a subset if 'columns' parameter is used.
@@ -524,7 +528,9 @@ def _arrow_to_geopandas(table, geo_metadata=None):
             )
 
     table_attr = table.drop(geometry_columns)
-    df = table_attr.to_pandas()
+    if to_pandas_kwargs is None:
+        to_pandas_kwargs = {}
+    df = table_attr.to_pandas(**to_pandas_kwargs)
 
     # Convert the WKB columns that are present back to geometry.
     for col in geometry_columns:
@@ -561,14 +567,7 @@ def _get_filesystem_path(path, filesystem=None, storage_options=None):
 
     If the filesystem is not None then it's just returned as is.
     """
-    import pyarrow
-
-    if (
-        isinstance(path, str)
-        and storage_options is None
-        and filesystem is None
-        and Version(pyarrow.__version__) >= Version("5.0.0")
-    ):
+    if isinstance(path, str) and storage_options is None and filesystem is None:
         # Use the native pyarrow filesystem if possible.
         try:
             from pyarrow.fs import FileSystem
@@ -587,7 +586,7 @@ def _get_filesystem_path(path, filesystem=None, storage_options=None):
 
     if filesystem is None and storage_options:
         raise ValueError(
-            "Cannot provide 'storage_options' with non-fsspec path '{}'".format(path)
+            f"Cannot provide 'storage_options' with non-fsspec path '{path}'"
         )
 
     return filesystem, path
@@ -668,7 +667,14 @@ def _read_parquet_schema_and_metadata(path, filesystem):
     return schema, metadata
 
 
-def _read_parquet(path, columns=None, storage_options=None, bbox=None, **kwargs):
+def _read_parquet(
+    path,
+    columns=None,
+    storage_options=None,
+    bbox=None,
+    to_pandas_kwargs=None,
+    **kwargs,
+):
     """
     Load a Parquet object from the file path, returning a GeoDataFrame.
 
@@ -716,6 +722,12 @@ def _read_parquet(path, columns=None, storage_options=None, bbox=None, **kwargs)
         Bounding box to be used to filter selection from geoparquet data. This
         is only usable if the data was saved with the bbox covering metadata.
         Input is of the tuple format (xmin, ymin, xmax, ymax).
+    to_pandas_kwargs : dict, optional
+        Arguments passed to the `pa.Table.to_pandas` method for non-geometry columns.
+        This can be used to control the behavior of the conversion of the non-geometry
+        columns to a pandas DataFrame. For example, you can use this to control the
+        dtype conversion of the columns. By default, the `to_pandas` method is called
+        with no additional arguments.
 
     **kwargs
         Any additional kwargs passed to :func:`pyarrow.parquet.read_table`.
@@ -751,6 +763,12 @@ def _read_parquet(path, columns=None, storage_options=None, bbox=None, **kwargs)
     schema, metadata = _read_parquet_schema_and_metadata(path, filesystem)
 
     geo_metadata = _validate_and_decode_metadata(metadata)
+    if len(geo_metadata["columns"]) == 0:
+        raise ValueError(
+            """No geometry columns are included in the columns read from
+            the Parquet/Feather file.  To read this file without geometry columns,
+            use pandas.read_parquet/read_feather() instead."""
+        )
 
     bbox_filter = (
         _get_parquet_bbox_filter(geo_metadata, bbox) if bbox is not None else None
@@ -776,10 +794,10 @@ def _read_parquet(path, columns=None, storage_options=None, bbox=None, **kwargs)
         path, columns=columns, filesystem=filesystem, filters=filters, **kwargs
     )
 
-    return _arrow_to_geopandas(table, geo_metadata)
+    return _arrow_to_geopandas(table, geo_metadata, to_pandas_kwargs)
 
 
-def _read_feather(path, columns=None, **kwargs):
+def _read_feather(path, columns=None, to_pandas_kwargs=None, **kwargs):
     """
     Load a Feather object from the file path, returning a GeoDataFrame.
 
@@ -793,7 +811,7 @@ def _read_feather(path, columns=None, **kwargs):
       columns, the first available geometry column will be set as the geometry
       column of the returned GeoDataFrame.
 
-    Supports versions 0.1.0, 0.4.0 and 1.0.0 of the GeoParquet
+    Supports versions 0.1.0, 0.4.0, 1.0.0 and 1.1.0 of the GeoParquet
     specification at: https://github.com/opengeospatial/geoparquet
 
     If 'crs' key is not present in the Feather metadata associated with the
@@ -812,6 +830,12 @@ def _read_feather(path, columns=None, **kwargs):
         geometry read from the file will be set as the geometry column
         of the returned GeoDataFrame.  If no geometry columns are present,
         a ``ValueError`` will be raised.
+    to_pandas_kwargs : dict, optional
+        Arguments passed to the `pa.Table.to_pandas` method for non-geometry columns.
+        This can be used to control the behavior of the conversion of the non-geometry
+        columns to a pandas DataFrame. For example, you can use this to control the
+        dtype conversion of the columns. By default, the `to_pandas` method is called
+        with no additional arguments.
     **kwargs
         Any additional kwargs passed to pyarrow.feather.read_table().
 
@@ -834,19 +858,12 @@ def _read_feather(path, columns=None, **kwargs):
     feather = import_optional_dependency(
         "pyarrow.feather", extra="pyarrow is required for Feather support."
     )
-    # TODO move this into `import_optional_dependency`
-    import pyarrow
-
     import geopandas.io._pyarrow_hotfix  # noqa: F401
-
-    if Version(pyarrow.__version__) < Version("0.17.0"):
-        raise ImportError("pyarrow >= 0.17 required for Feather support")
 
     path = _expand_user(path)
 
     table = feather.read_table(path, columns=columns, **kwargs)
-    _validate_and_decode_metadata(table.schema.metadata)
-    return _arrow_to_geopandas(table)
+    return _arrow_to_geopandas(table, to_pandas_kwargs=to_pandas_kwargs)
 
 
 def _get_parquet_bbox_filter(geo_metadata, bbox):
@@ -895,7 +912,6 @@ def _get_bbox_encoding_column_name(geo_metadata):
 
 
 def _get_non_bbox_columns(schema, geo_metadata):
-
     bbox_column_name = _get_bbox_encoding_column_name(geo_metadata)
     columns = schema.names
     if bbox_column_name in columns:
