@@ -172,11 +172,6 @@ html_theme_options = {
             "url": "https://github.com/geopandas/geopandas",
             "icon": "fab fa-github-square fa-xl",
         },
-        {
-            "name": "Twitter",
-            "url": "https://twitter.com/geopandas",
-            "icon": "fab fa-twitter-square fa-xl",
-        },
     ]
 }
 
@@ -245,6 +240,13 @@ moved_pages = [
 html_additional_pages = {page[0]: "redirect.html" for page in moved_pages}
 
 html_context = {"redirects": {old: new for old, new in moved_pages}}
+
+# Define the canonical URL if you are using a custom domain on Read the Docs
+html_baseurl = os.environ.get("READTHEDOCS_CANONICAL_URL", "")
+
+# Tell Jinja2 templates the build is running on Read the Docs
+if os.environ.get("READTHEDOCS", "") == "True":
+    html_context["READTHEDOCS"] = True
 
 # If false, no module index is generated.
 # html_domain_indices = True
