@@ -22,6 +22,7 @@ from matplotlib import cm, colors
 
 BRANCA_05 = Version(branca.__version__) > Version("0.4.2")
 FOLIUM_G_014 = Version(folium.__version__) > Version("0.14.0")
+FOLIUM_GE_019 = Version(folium.__version__) >= Version("0.19.0")
 
 
 @pytest.fixture(scope="class")
@@ -70,7 +71,10 @@ class TestExplore:
             pytest.approx(2.842170943040401e-14, rel=1e-6),
         ]
         assert m.options["zoom"] == 10
-        assert m.options["zoomControl"] is True
+        if not FOLIUM_GE_019:
+            assert m.options["zoomControl"] is True
+        else:
+            assert m.options["zoom_control"] is True
         assert m.position == "relative"
         assert m.height == (100.0, "%")
         assert m.width == (100.0, "%")
@@ -93,7 +97,10 @@ class TestExplore:
             pytest.approx(-73.9778006856748, rel=1e-6),
         ]
         assert m.options["zoom"] == 10
-        assert m.options["zoomControl"] is False
+        if not FOLIUM_GE_019:
+            assert m.options["zoomControl"] is False
+        else:
+            assert m.options["zoom_control"] is False
         assert m.height == (200.0, "px")
         assert m.width == (200.0, "px")
 
@@ -312,22 +319,30 @@ class TestExplore:
         assert '"__folium_color":"#9edae5","datetime":"2025-01-0101:22:00"' in out1_str
         assert '"__folium_color":"#1f77b4","datetime":"2022-01-0101:22:00"' in out1_str
 
+        df2 = df.set_index("datetime")
+        m2 = df2.explore()
+        out2_str = self._fetch_map_string(m2)
+        assert '"datetime":"2025-01-0101:22:00"' in out2_str
+        assert '"datetime":"2022-01-0101:22:00"' in out2_str
+
     def test_non_json_serialisable(self):
         df = self.nybb.copy().head(2)
-        uuid1 = uuid.UUID("12345678123456781234567812345678")
-        uuid2 = uuid.UUID("12345678123456781234567812345679")
+
+        u1 = "12345678-1234-5678-1234-567812345678"
+        uuid1 = uuid.UUID(u1)
+        u2 = "12345678-1234-5678-1234-567812345679"
+        uuid2 = uuid.UUID(u2)
         df["object"] = [uuid1, uuid2]
         m1 = df.explore("object")
 
         out1_str = self._fetch_map_string(m1)
-        assert (
-            '"__folium_color":"#9edae5","object":"12345678-1234-5678-1234-567812345679"}'
-            in out1_str
-        )
-        assert (
-            '"__folium_color":"#1f77b4","object":"12345678-1234-5678-1234-567812345678"'
-            in out1_str
-        )
+        assert f'"__folium_color":"#9edae5","object":"{u2}"' in out1_str
+        assert f'"__folium_color":"#1f77b4","object":"{u1}"' in out1_str
+        df2 = df.set_index("object")
+        m2 = df2.explore()
+        out2_str = self._fetch_map_string(m2)
+        assert f'"object":"{u2}"' in out2_str
+        assert f'"object":"{u1}"' in out2_str
 
     def test_string(self):
         df = self.nybb.copy()
@@ -371,7 +386,7 @@ class TestExplore:
             in out2_fields_str
         )
 
-        # GeoDataframe and the given list have different number of rows
+        # GeoDataFrame and the given list have different number of rows
         with pytest.raises(ValueError, match="different number of rows"):
             self.world.explore(column=np.array([1, 2, 3]))
 
@@ -743,7 +758,7 @@ class TestExplore:
             'attribution":"\\u0026copy;\\u003cahref=\\"https://www.openstreetmap.org'
             in out_str
         )
-        assert '"maxZoom":20,"minZoom":0' in out_str
+        assert '"maxZoom":20' in out_str
 
     @pytest.mark.skipif(not HAS_PYPROJ, reason="requires pyproj")
     def test_xyzservices_query_name(self):
@@ -760,7 +775,7 @@ class TestExplore:
             'attribution":"\\u0026copy;\\u003cahref=\\"https://www.openstreetmap.org'
             in out_str
         )
-        assert '"maxZoom":20,"minZoom":0' in out_str
+        assert '"maxZoom":20' in out_str
 
     @pytest.mark.skipif(not HAS_PYPROJ, reason="requires pyproj")
     def test_xyzservices_providers_min_zoom_override(self):
@@ -771,7 +786,7 @@ class TestExplore:
         )
         out_str = self._fetch_map_string(m)
 
-        assert '"maxZoom":20,"minZoom":3' in out_str
+        assert '"minZoom":3' in out_str
 
     @pytest.mark.skipif(not HAS_PYPROJ, reason="requires pyproj")
     def test_xyzservices_providers_max_zoom_override(self):
@@ -782,7 +797,7 @@ class TestExplore:
         )
         out_str = self._fetch_map_string(m)
 
-        assert '"maxZoom":12,"minZoom":0' in out_str
+        assert '"maxZoom":12' in out_str
 
     @pytest.mark.skipif(not HAS_PYPROJ, reason="requires pyproj")
     def test_xyzservices_providers_both_zooms_override(self):
@@ -795,7 +810,8 @@ class TestExplore:
         )
         out_str = self._fetch_map_string(m)
 
-        assert '"maxZoom":12,"minZoom":3' in out_str
+        assert '"maxZoom":12' in out_str
+        assert '"minZoom":3' in out_str
 
     def test_linearrings(self):
         rings = self.nybb.explode(index_parts=True).exterior
@@ -971,9 +987,15 @@ class TestExplore:
     def test_map_kwds(self):
         def check():
             out_str = self._fetch_map_string(m)
-            assert "zoomControl:false" in out_str
-            assert "dragging:false" in out_str
-            assert "scrollWheelZoom:false" in out_str
+            if not FOLIUM_GE_019:
+                assert m.options["zoomControl"] is False
+            else:
+                assert m.options["zoom_control"] is False
+            assert "dragging:false" in out_str or '"dragging":false' in out_str
+            assert (
+                "scrollWheelZoom:false" in out_str
+                or '"scrollWheelZoom":false' in out_str
+            )
 
         # check that folium and leaflet Map() parameters can be passed
         m = self.world.explore(
