@@ -5,11 +5,11 @@ from .sjoin import sjoin, sjoin_nearest
 from .util import collect
 
 __all__ = [
+    "clip",
     "collect",
     "geocode",
     "overlay",
     "reverse_geocode",
     "sjoin",
     "sjoin_nearest",
-    "clip",
 ]
